+++ griptape/memory/meta/__init__.py
@@ -1,5 +1,5 @@
 from .base_meta_entry import BaseMetaEntry
-from .base_meta_memory import MetaMemory
+from .meta_memory import MetaMemory
 
 __all__ = [
     "BaseMetaEntry",
